Changed spec of Postgres Kind
- Renamed setupCommands to commands
- Added more fields to the Postgres Status
- Refactored the controller code

It is now possible to perform actions on the same
Postgres instance multiple times. For example, you
may want setup the Postgres instance with some database
at when the instance is created. Later you may want to
add additional databases to the same instance. This is
now possible.

Check artifacts/examples/client10-postgres.yaml and
artifacts/examples/client10-postgres-update.yaml to see
this aspect of the controller/operator.