Added version 2 of the Postgres operator
This version supports declarative creation of users and databases
The Operator performs diffs of the desired state and the current
state and decides whether to add/modify/delete entities
(users and databases).

Additionally, other Postgres commands are supported as well.
For this a generic "commands" attribute is defined in the CRD spec.
Using this attribute one can specify any Postgres command.
This aspect of the Spec deviates from the purely declarative format.

The combination of declarative approach for handling user and
database commands and imperative approach for all other commands has
been done with intention that the common commands should be
made simple, while other commands should be possible.