WIP openapi-spec updates -- DO NOT Merge
Modified build-openapi-spec.sh to support dirname
and filename as inputs. There seem to be some Operators
where Custom resource types are not defined in types.go
but in some custom go file. This change in build-openapi-spec.sh
supports such cases. We need more design and thought around
how to handle such cases in a general manner.
(We saw this in Oracle's MySQL Operator).

Another thing we saw in that Operator is types definition
file also contained certain function definitions.
build-openapi-spec.sh needs to be able to work with such files as well.