README updates and added placholder DeleteFunc
The purpose of DeleteFunc is to experiment with
'ordered' delete of Objects - Deployment and Service.
The ordering does not really matter this CRD, but
there was a thread on kubernetes-dev[1] about how
to perform ordered deletes. Our view is that just
OwnerReferences does not help with the order. Required
order needs to be enforced in the delete handler.
The placeholder delete handler needs to be fleshed
out to implement order deletes. Note that the DeleteFunc
EventHandlerSetup is commented out right now.
Once the delete handler is fleshed out, uncomment it.

Right now without the delete handler, Kubernetes correctly
deletes all the child Objects when Postgres instance is deleted.
Only that it does not do this in any specific order.

[1] https://groups.google.com/forum/#!topic/kubernetes-dev/1QnYSPGn85M