Removed dependency on operator-discovery-helper
operator-discovery-helper module was written as a bridge
between kubediscovery and the crds available in the cluster.
operator-discovery-helper 'discovered' available crds and
shared them with kubediscovery through etcd (different from
the control plane's own etcd).

kubediscovery has now been updated to list crds itself by grafting
the appropriate operator-discovery-helper code inside it.

We are not yet removing etcd container deployment as part of rc.yaml
as it can be used for storing the composition graph.
If we decide to not store the composition graph then we can remove
etcd dependency as well. But that will be in a separate commit.