Example to demonstrate team-level multitenancy
In this example Platform Engineering team registers
a Tenant Service that defines template for creating
Namespace, Service Account and specifying the resource
limits that should be defined on any Pod that is deployed
in that Namespace. The policy also defines that each Pod
should have cpu and memory requests defined and that their
values can be different for different tenants as needed
by the tenants.

Product teams create workloads in their tenant's Namespace later.
KubePlus ensures that the workload Pods are provided with
appropriate cpu/memory requests and limits values.