Creating service instances in separate namespaces
B2B SaaS on K8s is achieved by deploying separate instance
of the application for each customer. To achieve strict
isolation and multi-tenancy, such instances are deployed in
separate namespaces. This is called as the multi-instance
multi-tenancy of the K8S based SaaS model [1].

This patch adds the support to enable this SaaS model in KubePlus.
KubePlus will now create a new namespace for deploying the
Helm release corresponding to a service instance.
The name of the namespace is taken to be same as the name
of the service instance. KubePlus first creates the NS and then passes
that to the Helm install command. All the resources that
are part of the Helm release are created in this NS.
Note that following restrictions are required on the Helm chart
underlying the service.
- The Helm chart should not contain Namespace definition. This
is because all the Helm chart resources are going to be created in the new
NS that KubePlus creates for that instance. Since Kubernetes Namespaces cannot be nested
the Namespace defined in the Helm chart will get created just as an object
in the NS that KubePlus creates. But it won't contain any of the actual service instance
resources. They will be present in the NS that KubePlus has created.
- The Helm chart should not take Namespace field as input through
values.yaml. This is because the Namespace name so entered will
be overridden by the NS that KubePlus creates.
We are tracking above two points in below issues:

https://github.com/cloud-ark/kubeplus/issues/910
https://github.com/cloud-ark/kubeplus/issues/911

[1] https://cloudark.medium.com/multi-instance-multi-tenancy-in-kubernetes-273cd22ae019