Denies deletes to starting app instances and async delete.
When a new application instance is created, it takes some time for it to start up and update its status. This change prevents the deletion of the instance during this phase to make sure that the delete operation returns an error message when the instance can't be deleted. This commit also makes the delete operation asynchronous since `helm delete` can take a while, especially when there are many instances deleted at once.

Resolves: #1178
Resolves: #1180