For metrics and logs using Pods within the NS of the instance
In the original design, we were using the output of kubectl connections
to find out the Pods and then running metrics and logs on those Pods.
However, for workloads in which Pods get created at runtime, connections
is not able to find all the newly created workload Pods. This leads to
incomplete results for metrics and logs. A simple way to handle this is
to use all the pods in the namespace in which the application pods are running
for collecting logs and metrics. The runtime Pods are typically
created in the same namespace in which the application is running.
This will give accurate results (atleast more accurate than using the output
of kubect connections for the list of Pods).

Fixes: https://github.com/cloud-ark/kubeplus/issues/1190