Metrics plugin update (#1408)
When collecting Pods in a namespace for an app instance, we want 
to ensure that the namespace is indeed for the Kind whose app instance
is being passed to the kubectl metrics command.
Without this Kind name check, kubeplus may get wrong Pods if
app instances for two different Kinds are given the same name.